% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmf.R
\name{pmf}
\alias{pmf}
\title{This function is going to generate the + insted of a TRUE value}
\usage{
pmf(x)
}
\arguments{
\item{x}{Include the vector to be tranformed}
}
\value{
a vector of + and - values
}
\description{
Use it for table generation where T or F values need to be nicely presented
}
